Fixed some warnings
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@703 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/http_headers.h
@@ -90,8 +90,8 @@ class http_headers
 public:
 	static int buildHTTPRequestHeaderStruct(HTTP_REQUEST_HEADER *request,httpThreadContext *td,char *input=0);
 	static int buildHTTPResponseHeaderStruct(HTTP_RESPONSE_HEADER *response,httpThreadContext *td,char *input=0);
-	static u_long validHTTPRequest(char*,httpThreadContext*,u_long*,u_long*);
-	static u_long validHTTPResponse(char*,httpThreadContext*,u_long*,u_long*);
+	static int validHTTPRequest(char*,httpThreadContext*,u_long*,u_long*);
+	static int validHTTPResponse(char*,httpThreadContext*,u_long*,u_long*);
 	static void resetHTTPRequest(HTTP_REQUEST_HEADER *request);
 	static void resetHTTPResponse(HTTP_RESPONSE_HEADER *response);
 	static void buildDefaultHTTPResponseHeader(HTTP_RESPONSE_HEADER*);
